refactor to fix eslint errors

+++ src/utils/funcs/check-table.ts
@@ -1,4 +1,4 @@
-import Database from "@tauri-apps/plugin-sql";
+import Database from '@tauri-apps/plugin-sql';
 
 export const checkTable = (
   db: Database,
@@ -6,23 +6,21 @@ export const checkTable = (
   createFunc: (db: Database) => void,
   strings: string[],
   ints: string[],
-  reals?: string[]
+  reals?: string[],
 ) => {
   db.select(`PRAGMA table_info(${table})`)
-    .then((res: any) => {
-      if (res && res.length) {
-        if (
-          res.length !==
-          strings.length + ints.length + (reals ? reals.length : 0)
-        ) {
+    .then(res => {
+      if (res && typeof res === 'object' && 'length' in res && typeof res.length === 'number') {
+        const checkedRes = res as { cid: number; name: string; type: string; notnull: number }[];
+        if (res.length !== strings.length + ints.length + (reals ? reals.length : 0)) {
           db.execute(`DROP table ${table}`).then(() => createFunc(db));
           return;
         } else {
-          res.forEach((element: any) => {
+          checkedRes.forEach(element => {
             if (
-              !(strings.includes(element.name) && element.type === "TEXT") &&
-              !(ints.includes(element.name) && element.type === "INTEGER") &&
-              !(reals?.includes(element.name) && element.type === "REAL")
+              !(strings.includes(element.name) && element.type === 'TEXT') &&
+              !(ints.includes(element.name) && element.type === 'INTEGER') &&
+              !(reals?.includes(element.name) && element.type === 'REAL')
             ) {
               db.execute(`DROP table ${table}`).then(() => createFunc(db));
               return false;
@@ -33,7 +31,7 @@ export const checkTable = (
         createFunc(db);
       }
     })
-    .catch((err) => {
+    .catch(err => {
       console.log(err);
     });
 };
